debug skip task

+++ src/commands/Minion/slayertask.ts
@@ -9,6 +9,7 @@ import {
 } from '../../lib/slayer/slayerUtil';
 import { BotCommand } from '../../lib/structures/BotCommand';
 import { stringMatches } from '../../lib/util';
+import { Monsters } from 'oldschooljs';
 
 export default class extends BotCommand {
 	public constructor(store: CommandStore, file: string[], directory: string) {
@@ -54,24 +55,26 @@ export default class extends BotCommand {
 			? slayerMasters.find(m => m.aliases.some(alias => stringMatches(alias, input))) ?? null
 			: null;
 
-		if (currentTask || !slayerMaster) {
-			let warningInfo = '';
-			if (input && !slayerMaster && matchedSlayerMaster) {
-				warningInfo = `You do not have the requirements to use ${matchedSlayerMaster.name}.\n\n`;
-			}
-			let baseInfo = currentTask
-				? `Your current task is to kill ${currentTask.quantity}x ${
+		if (!msg.flagArgs.skip && (currentTask || !slayerMaster)) {
+		//if (currentTask || !slayerMaster) {
+				let warningInfo = '';
+				if (input && !slayerMaster && matchedSlayerMaster) {
+					warningInfo = `You do not have the requirements to use ${matchedSlayerMaster.name}.\n\n`;
+				}
+				let baseInfo = currentTask
+					? `Your current task is to kill ${currentTask.quantity}x ${
 						assignedTask!.monster.name
-				  }, you have ${currentTask.quantityRemaining} kills remaining.`
-				: `You have no task at the moment <:FrogBigEyes:847859910933741628> You can get a task using \`${
+					}, you have ${currentTask.quantityRemaining} kills remaining.`
+					: `You have no task at the moment <:FrogBigEyes:847859910933741628> You can get a task using \`${
 						msg.cmdPrefix
-				  }slayertask ${slayerMasters.map(i => i.name).join('/')}}\``;
+					}slayertask ${slayerMasters.map(i => i.name).join('/')}}\``;
 
-			return msg.channel.send(`${warningInfo}${baseInfo}
+				return msg.channel.send(`${warningInfo}${baseInfo}
 		
 You've done ${totalTasksDone} tasks. Your current streak is ${msg.author.settings.get(
-				UserSettings.Slayer.TaskStreak
-			)}.`);
+					UserSettings.Slayer.TaskStreak
+				)}.`);
+			}
 		}
 
 		// Store favorite slayer master if requested:
@@ -80,6 +83,23 @@ You've done ${totalTasksDone} tasks. Your current streak is ${msg.author.setting
 		}
 
 		const newSlayerTask = await assignNewSlayerTask(msg.author, slayerMaster);
+		let commonName = newSlayerTask.assignedTask.monster.name;
+		switch (newSlayerTask.assignedTask.monster.id) {
+			case Monsters.KaphiteWorker:
+				commonName = 'Kalphite';
+				break;
+			case Monsters.MountainTroll:
+				commonName = 'Trolls';
+				break;
+			case Monsters.FossilIslandWyvernSpitting:
+				commonName = 'Fossil Island Wyverns';
+				break;
+			case Monsters.FeralVampyre:
+				commonName = 'Vampyres';
+				break;
+
+			default:
+		}
 		return msg.channel.send(
 			`${slayerMaster.name} has assigned you to kill ${newSlayerTask.currentTask.quantity}x ${newSlayerTask.assignedTask.monster.name}.`
 		);
